+++ src/components/layout/admin/nav/Sidebar.tsx
@@ -1,8 +1,8 @@
 import Button from '@/components/common/Button';
 import QrCodeModal from '@/components/qrCodeModal/QrCodeModal';
 import { modeSubText, modeText } from '@/data/admin';
-import { useAuthSetQuery } from '@/hooks/auth/useAuthSetQuery';
 import { useModal } from '@/hooks/modal/useModal';
+import { useAuthSetQuery } from '@/hooks/query/auth/useAuthSetQuery';
 import useSideBar, { setIsSideBarOpen } from '@/shared/store/sidebar';
 import useToggleState from '@/shared/store/toggle';
 import clsx from 'clsx';
