+++ src/hooks/query/auth/useAuthSetQuery.ts
@@ -14,8 +14,8 @@ import useAuthState, { setSession, setStoreBno, setStoreId, setStoreName } from 
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { useRouter } from 'next/router';
 import { useState } from 'react';
-import { useModal } from '../modal/useModal';
-import useToast from '../toast/useToast';
+import { useModal } from '../../modal/useModal';
+import useToast from '../../toast/useToast';
 
 const enum QUERY_KEY {
   LOGIN = 'login',
